Commandã 객체로 Strategy Pattern 적용 및 그로인한 Mock Test는 불가능하여 5월8일까지 Mock Test Code 유효
