--- src/main/java/ssd/SSDInterface.java
@@ -1,7 +1,0 @@
-package ssd;
-
-public interface SSDInterface {
-    void read(Command command);
-    void write(Command command);
-    void erase(Command command);
-}
